+++ packages/ui/src/hero-sections.tsx
@@ -1,11 +1,11 @@
 import { ReactNode } from 'react'
-import { IconLogoGeme,IconCycleArrow } from './icons'
+import { IconLogoGeme } from './icons'
 // import { IVideoPlayerProps, VideoPlayer } from './video'
 import { IMuxVideoPlayerProps, MuxVideoPlayer } from './mux-video'
 import { ILinkComponent } from '../../../apps/gemebuild/src/contexts/link-context'
 import Image from 'next/image'
 import VideoInlinePlayer, { PlayIcon } from './video-inline-player'
-import mixpanel from 'mixpanel-browser'
+import HeroSectionsButton from './client-components/HeroSectionsButton'
 
 interface IHeroSection1Props {
   title?: ReactNode
@@ -71,12 +71,6 @@ function HeroSection1({
   LinkComponent,
   fullScreenVideoUrl,
 }: IHeroSection1Props) {
-  const linkClickedTrack = () => {
-    mixpanel.track('Go Product Page', {
-      From: 'Hero Primary Button',
-    })
-  }
-  
   return (
     <div className="h-screen ">
       {/* 大屏图片层 */}
@@ -103,27 +97,11 @@ function HeroSection1({
             {description}
           </p>
         )}
-        {linkText && linkUrl && LinkComponent == null && (
-        <a
-          className=" v2311-font-h3 font-bold text-white xl:w-auto flex items-center justify-center mt-[80px] md:mt-[160px] xl:mt-16 border-solid border-2 border-white rounded-xl py-1 md:py-2 px-16 md:px-32 xl:px-8"
-          href={linkUrl}
-          onClick={linkClickedTrack}
-        >
-          {linkText}
-          <IconCycleArrow className="inline-block ml-4 md:ml-8 xl:ml-4 w-5 md:w-10 xl:w-12 h-5 md:h-10 xl:h-12" />
-        </a>
-      )}
-      {linkText && linkUrl && LinkComponent && (
-        <span onClick={linkClickedTrack}>
-          <LinkComponent
-            className=" v2311-font-h3 font-bold text-white xl:w-auto flex items-center justify-center mt-[80px] md:mt-[160px] xl:mt-16 border-solid border-2 border-white rounded-xl py-1 md:py-2 px-16 md:px-32 xl:px-8"
-            href={linkUrl}
-          >
-            {linkText}
-            <IconCycleArrow className="inline-block ml-4 md:ml-8 xl:ml-4 w-5 md:w-10 xl:w-12 h-5 md:h-10 xl:h-12" />
-          </LinkComponent>
-        </span>
-      )}
+        <HeroSectionsButton
+          LinkComponent={LinkComponent}
+          linkText={linkText}
+          linkUrl={linkUrl}
+        />
       </div>
 
       {/* 全屏播放器 */}
